[client] extract `send_and_wait_for_in_block`
