+++ client/src/utils.rs
@@ -1,10 +1,12 @@
+use crate::exit_code;
 use codec::Encode;
 use encointer_primitives::scheduler::CeremonyIndexType;
+use log::{debug, error, info};
 use sp_application_crypto::sr25519;
-use sp_core::Pair;
+use sp_core::{Pair, H256};
 use substrate_api_client::{
 	compose_call, compose_extrinsic, compose_extrinsic_offline, rpc::WsRpcClient, Api, Metadata,
-	UncheckedExtrinsicV4,
+	UncheckedExtrinsicV4, XtStatus,
 };
 
 /// Wrapper around the `compose_extrinsic_offline!` macro to be less verbose.
@@ -40,6 +42,40 @@ pub fn sudo_xt<C: Encode + Clone>(
 /// Wraps the supplied calls in a batch call
 pub fn batch_call<C: Encode + Clone>(metadata: &Metadata, calls: Vec<C>) -> ([u8; 2], Vec<C>) {
 	compose_call!(metadata, "Utility", "batch", calls)
+}
+
+pub fn send_and_wait_for_in_block<C: Encode>(
+	api: &Api<sr25519::Pair, WsRpcClient>,
+	xt: UncheckedExtrinsicV4<C>,
+) -> Option<H256> {
+	ensure_payment(&api, &xt.hex_encode());
+	let tx_hash = api.send_extrinsic(xt.hex_encode(), XtStatus::InBlock).unwrap();
+	info!("[+] Transaction got included. Hash: {:?}\n", tx_hash);
+
+	tx_hash
+}
+
+/// Checks if the account has sufficient funds. Exits the process if not.
+pub fn ensure_payment(api: &Api<sr25519::Pair, WsRpcClient>, xt: &str) {
+	let signer_balance = match api.get_account_data(&api.signer_account().unwrap()).unwrap() {
+		Some(bal) => bal.free,
+		None => {
+			error!("account does not exist on chain");
+			std::process::exit(exit_code::FEE_PAYMENT_FAILED);
+		},
+	};
+	let fee = api
+		.get_fee_details(xt, None)
+		.unwrap()
+		.unwrap()
+		.inclusion_fee
+		.map_or_else(|| 0, |details| details.base_fee);
+	let ed = api.get_existential_deposit().unwrap();
+	if signer_balance < fee + ed {
+		error!("insufficient funds: fee: {} ed: {} bal: {:?}", fee, ed, signer_balance);
+		std::process::exit(exit_code::FEE_PAYMENT_FAILED);
+	}
+	debug!("account can pay fees: fee: {} ed: {} bal: {}", fee, ed, signer_balance);
 }
 
 /// Handles the potential case of a negative ceremony index CLI.
